LPS-170922 Moves to the module where we are using it
