+++ modules/apps/site/site-memberships-web/src/main/java/com/liferay/site/memberships/web/internal/search/UserGroupSiteMembershipChecker.java
@@ -12,23 +12,23 @@
  * details.
  */
 
-package com.liferay.portlet.sitesadmin.search;
+package com.liferay.site.memberships.web.internal.search;
 
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.Organization;
-import com.liferay.portal.kernel.service.OrganizationLocalServiceUtil;
+import com.liferay.portal.kernel.model.UserGroup;
+import com.liferay.portal.kernel.service.UserGroupLocalServiceUtil;
 
 import javax.portlet.RenderResponse;
 
 /**
  * @author Charles May
  */
-public class OrganizationSiteMembershipChecker extends EmptyOnClickRowChecker {
+public class UserGroupSiteMembershipChecker extends EmptyOnClickRowChecker {
 
-	public OrganizationSiteMembershipChecker(
+	public UserGroupSiteMembershipChecker(
 		RenderResponse renderResponse, Group group) {
 
 		super(renderResponse);
@@ -38,37 +38,26 @@ public class OrganizationSiteMembershipChecker extends EmptyOnClickRowChecker {
 
 	@Override
 	public boolean isChecked(Object object) {
-		Organization organization = (Organization)object;
+		UserGroup userGroup = (UserGroup)object;
 
 		try {
-			if (OrganizationLocalServiceUtil.hasGroupOrganization(
-					_group.getGroupId(), organization.getOrganizationId()) ||
-				(_group.getOrganizationId() ==
-					organization.getOrganizationId())) {
-
-				return true;
-			}
+			return UserGroupLocalServiceUtil.hasGroupUserGroup(
+				_group.getGroupId(), userGroup.getUserGroupId());
 		}
 		catch (Exception exception) {
 			_log.error(exception);
-		}
 
-		return false;
+			return false;
+		}
 	}
 
 	@Override
 	public boolean isDisabled(Object object) {
-		Organization organization = (Organization)object;
-
-		if (_group.getOrganizationId() == organization.getOrganizationId()) {
-			return true;
-		}
-
 		return isChecked(object);
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		OrganizationSiteMembershipChecker.class);
+		UserGroupSiteMembershipChecker.class);
 
 	private final Group _group;
