+++ modules/apps/site/site-memberships-web/src/main/java/com/liferay/site/memberships/web/internal/search/UserSiteMembershipChecker.java
@@ -12,13 +12,14 @@
  * details.
  */
 
-package com.liferay.portlet.sitesadmin.search;
+package com.liferay.site.memberships.web.internal.search;
 
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Team;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.membershippolicy.SiteMembershipPolicyUtil;
 import com.liferay.portal.kernel.service.UserLocalServiceUtil;
 
 import javax.portlet.RenderResponse;
@@ -26,30 +27,33 @@ import javax.portlet.RenderResponse;
 /**
  * @author Brian Wing Shun Chan
  */
-public class UserTeamChecker extends EmptyOnClickRowChecker {
+public class UserSiteMembershipChecker extends EmptyOnClickRowChecker {
 
-	public UserTeamChecker(RenderResponse renderResponse, Team team) {
+	public UserSiteMembershipChecker(
+		RenderResponse renderResponse, Group group) {
+
 		super(renderResponse);
 
-		_team = team;
+		_group = group;
 	}
 
 	@Override
 	public boolean isChecked(Object object) {
-		return hasTeamUser(object);
-	}
+		User user = null;
 
-	@Override
-	public boolean isDisabled(Object object) {
-		return hasTeamUser(object);
-	}
-
-	protected boolean hasTeamUser(Object object) {
-		User user = (User)object;
+		if (object instanceof User) {
+			user = (User)object;
+		}
+		else if (object instanceof Object[]) {
+			user = (User)((Object[])object)[0];
+		}
+		else {
+			throw new IllegalArgumentException(object + " is not a user");
+		}
 
 		try {
-			return UserLocalServiceUtil.hasTeamUser(
-				_team.getTeamId(), user.getUserId());
+			return UserLocalServiceUtil.hasGroupUser(
+				_group.getGroupId(), user.getUserId());
 		}
 		catch (Exception exception) {
 			_log.error(exception);
@@ -58,9 +62,28 @@ public class UserTeamChecker extends EmptyOnClickRowChecker {
 		}
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(
-		UserTeamChecker.class);
+	@Override
+	public boolean isDisabled(Object object) {
+		User user = (User)object;
 
-	private final Team _team;
+		try {
+			if (isChecked(user) ||
+				!SiteMembershipPolicyUtil.isMembershipAllowed(
+					user.getUserId(), _group.getGroupId())) {
+
+				return true;
+			}
+		}
+		catch (Exception exception) {
+			_log.error(exception);
+		}
+
+		return super.isDisabled(object);
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		UserSiteMembershipChecker.class);
+
+	private final Group _group;
 
 }
