+++ modules/apps/site/site-teams-web/src/main/java/com/liferay/site/teams/web/internal/search/UserGroupTeamChecker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.sitesadmin.search;
+package com.liferay.site.teams.web.internal.search;
 
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.log.Log;
